refactor: Resolve redundant clone warnings and improve code quality
- Applied cargo clippy --fix to automatically resolve redundant clones and simplifiable patterns
- Manually optimized ner_engine.rs to use std::slice::from_ref instead of cloning PathBuf
- Reduced total clippy warnings from 27 to 7
- Improved consistency in iterator usage and simplified boolean logic

+++ src/summary/presentation.rs
@@ -165,7 +165,7 @@ impl EntitySummary {
         };
 
         let now = Utc::now();
-        let days_since_last_seen = (now - analysis.last_seen).num_days() as i64;
+        let days_since_last_seen = (now - analysis.last_seen).num_days();
         let recency_level = match days_since_last_seen {
             0 => RecencyLevel::Today,
             1..=3 => RecencyLevel::Recent,
